requirements-dev.txt
django>=4.0.5
djangorestframework
django-cors-headers
django-extensions
lorem-text
